refactor(portability): Changed vmid_t to ctx_id_t and further uint64_t substitutions
The rationale behind the change from vmid_t to ctx_id_t stems from the current usage of the parameters passed to tlb_vm_inv_va and tlb_vm_inv_all inline functions. These functions currently receives the vmid parameter from the field addr_space_t::id, which is defined as a uint64_t. The field addr_space_t::id does not identify VMs exclusively, but it can assume the value of 0 to identify the hypervisor address space. This change affects the associated type of addr_space_t::id and tlb_vm_inv_* functions. The new type ctx_id_t introduced in this commit represents a mean to identify the context of an object, be it tThe rationale behind the change from vmid_t to ctx_id_t stems from the current usage of the parameters passed to tlb_vm_inv_va and tlb_vm_inv_all inline functions. These functions currently receives the vmid parameter from the field addr_space_t::id, which is defined as a uint64_t. The field addr_space_t::id does not identify VMs exclusively, but it can assume the value of 0 to identify the hypervisor address space. This change affects the associated type of addr_space_t::id field and tlb_vm_inv_* functions. The new type ctx_id_t introduced in this commit represents a value which can be used to identify the context of an object, be it the hypervisor's context (id=0) or a VM's context (id>0).

Signed-off-by: Aras Ashraf Gandomi <aras.gandomi@gmail.com>